requirements.txt
CoAPthon3==1.0.1
redis==3.2.1
walrus==0.7.1
